Division into modules and correct of remarks
